renamed fixture to request_create_project
